Create ProjectInstanceExtensions per PR comments
